Attempt to encrypt plaintext passwords before attempting to verify them
This is a worthwhile modification to the verification logic which
should help ensure that plaintext passwords stay around in the
database for as little amount of time as possible.

It does have the slight disadvantage that this would then immediately
call back into the crypto module it just used to encrypt it in order to
verify it, which would have twice the computational overhead. However,
this would be a one-time cost for a single login attempt; after that, it
would be encrypted already, and then it's the same computational cost as
any other login attempt.

Reported-By: glguy@libera.chat